# docs/swagger.yaml
basePath: /
definitions:
...
  request.JournalEntryCreateRequest:
    properties:
      date:
        type: string
      description:
        minLength: 1
        type: string
      lines:
        items:
          $ref: '#/definitions/request.JournalLineRequest'
        minItems: 2
        type: array
    required:
    - date
    - description
    - lines
    type: object
  request.JournalEntryUpdateRequest:
    properties:
      date:
        type: string
      description:
        minLength: 1
        type: string
      lines:
        items:
          $ref: '#/definitions/request.JournalLineRequest'
        minItems: 2
        type: array
    required:
    - date
    - description
    - lines
    type: object
  request.JournalLineRequest:
    properties:
      coa_id:
        type: string
      credit:
        minimum: 0
        type: number
      debit:
        minimum: 0
        type: number
      description:
        minLength: 1
        type: string
    required:
    - coa_id
    - description
    type: object
...
  response.JSON:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      status:
        type: integer
    type: object
  response.Links:
    properties:
      current_page:
        type: string
      first_page:
        type: string
      last_page:
        type: string
      next_page:
        type: string
      prev_page:
        type: string
    type: object
  response.Meta:
    properties:
      info:
        type: string
      links:
        $ref: '#/definitions/response.Links'
      page:
        type: integer
      page_size:
        type: integer
      search:
        type: string
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
...
host: 127.0.0.1:3000
info:
  contact:
    email: fatihrizqon@gmail.com
    name: Fatih Rizqon
  description: This is an Official Documentation for Go REST API with Fiber Framework
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go REST API with Fiber Framework
  version: "1.0"
paths:
...
  /api/v1/journal_entries:
    get:
      parameters:
      - description: Search keyword
        in: query
        name: search
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Filter by status (draft, posted, void)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSON'
      summary: Get all general journal entries
      tags:
      - JournalEntries
    post:
      consumes:
      - application/json
      parameters:
      - description: Journal Entry Create Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.JournalEntryCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.JSON'
      security:
      - BearerAuth: []
      summary: Create general journal entry
      tags:
      - JournalEntries
  /api/v1/journal_entries/{id}:
    delete:
      parameters:
      - description: Journal Entry ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSON'
      summary: Delete journal entry
      tags:
      - JournalEntries
    get:
      parameters:
      - description: Journal Entry ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSON'
      summary: Get journal entry by ID
      tags:
      - JournalEntries
    put:
      consumes:
      - application/json
      parameters:
      - description: Journal Entry ID
        in: path
        name: id
        required: true
        type: string
      - description: Journal Entry Update Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.JournalEntryUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSON'
      security:
      - BearerAuth: []
      summary: Update journal entry
      tags:
      - JournalEntries
  /api/v1/journal_entries/{id}/post:
    put:
      parameters:
      - description: Journal Entry ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSON'
      summary: Post journal entry
      tags:
      - JournalEntries
  /api/v1/journal_entries/{id}/void:
    put:
      parameters:
      - description: Journal Entry ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSON'
      summary: Void journal entry
      tags:
      - JournalEntries
...
  /api/v1/revenues:
    get:
      parameters:
      - description: Search keyword
        in: query
        name: search
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSON'
      summary: Get all revenue entries
      tags:
      - Revenue
    post:
      consumes:
      - application/json
      description: Store a new revenue journal entry (type=revenue, prefix RV-).
      parameters:
      - description: Revenue Create Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.JournalEntryCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.JSON'
      security:
      - BearerAuth: []
      summary: Create revenue entry
      tags:
      - Revenue
  /api/v1/revenues/{id}:
    delete:
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSON'
      summary: Delete revenue entry
      tags:
      - Revenue
    get:
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSON'
      summary: Get revenue entry by ID
      tags:
      - Revenue
    put:
      consumes:
      - application/json
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: string
      - description: Revenue Update Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.JournalEntryUpdateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSON'
      security:
      - BearerAuth: []
      summary: Update revenue entry
      tags:
      - Revenue
  /api/v1/revenues/{id}/post:
    put:
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSON'
      summary: Post revenue entry
      tags:
      - Revenue
  /api/v1/revenues/{id}/void:
    put:
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSON'
      summary: Void revenue entry
      tags:
      - Revenue
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"